py/requirements.txt
--index-url https://pypi.devinfra.sentry.io/simple

sentry-infra-event-notifier==0.0.4
